feat: set up basic portfolio app that embeds the wallet panel
