+++ .changeset/moody-squids-repair.md
@@ -1,0 +1,5 @@
+---
+"@fogo/sessions-sdk-react": patch
+---
+
+Separate SessionPanel component and export it
