+++ apps/paymaster-admin/package.json
@@ -6,8 +6,8 @@
     "build": "next build",
     "fix:format": "prettier --write .",
     "fix:lint": "eslint --fix .",
-    "start:dev": "next dev --port 3001",
-    "start:prod": "next start --port 3001",
+    "start:dev": "next dev --port 3002",
+    "start:prod": "next start --port 3002",
     "test:format": "prettier --check .",
     "test:lint": "eslint .",
     "test:types": "tsc"
