+++ apps/portfolio/build.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+set -e
+
+. "$HOME/.cargo/env"
+cd ../..
+turbo run build --filter={apps/portfolio}...
