+++ apps/portfolio/src/components/Root/index.module.scss
@@ -1,168 +1,5 @@
-@use "../../theme";
-
-.root {
-  // We use `scrollbar-gutter: stable` which prevents the page from jumping when
-  // adding or removing the scrollbar.  However, react-aria [tries to add a
-  // padding](https://github.com/adobe/react-spectrum/issues/5470) to the html
-  // element when opening/closing popovers and does not account for
-  // `scrollbar-gutter`, and there's no way (yet) to disable that behavior.
-  // Forcing the padding to zero here effectively prevents that behavior from
-  // causing the page to jump.
-  // TODO Remove this when a fix for
-  // https://github.com/adobe/react-spectrum/issues/5470 lands in react-aria
-  scrollbar-gutter: stable;
-  padding-right: 0 !important;
-  overflow-x: clip;
-
-  @include theme.max-width-set-breakpoints;
-
-  .body {
-    overflow-x: clip;
-    display: grid;
-    grid-template-rows: max-content max-content 1fr max-content;
-    grid-template-columns: 100%;
-    min-height: 100lvh;
-
-    @include theme.breakpoint("lg") {
-      grid-template-columns: 1fr theme.$max-width 1fr;
-      grid-template-rows: 1fr max-content;
-    }
-
-    .header {
-      position: sticky;
-      top: 0;
-
-      @include theme.breakpoint("lg") {
-        display: none;
-      }
-    }
-
-    .sidebar {
-      display: none;
-      grid-template-rows: max-content 1fr max-content;
-      z-index: 1;
-      height: 100dvh;
-      position: sticky;
-      top: 0;
-
-      @include theme.breakpoint("lg") {
-        display: grid;
-        grid-row: span 2 / span 2;
-      }
-
-      @include theme.breakpoint("xl") {
-        width: theme.spacing(60);
-        justify-self: end;
-      }
-
-      .logoContainer {
-        padding: theme.spacing(6);
-        display: flex;
-      }
-
-      .navButtons {
-        display: flex;
-        flex-flow: column nowrap;
-        justify-content: center;
-        gap: theme.spacing(8);
-        align-items: stretch;
-
-        .navGroup {
-          display: flex;
-          flex-flow: column nowrap;
-          justify-content: center;
-          gap: theme.spacing(1);
-          align-items: stretch;
-        }
-      }
-
-      .sidebarSocialLinks {
-        padding: theme.spacing(6);
-        display: flex;
-        flex-flow: row nowrap;
-        gap: theme.spacing(4);
-      }
-    }
-
-    footer {
-      text-align: center;
-      padding-bottom: theme.spacing(6);
-
-      @include theme.max-width;
-
-      @include theme.breakpoint("lg") {
-        grid-row: 2;
-        grid-column: 2;
-        width: 100%;
-      }
-
-      .lines {
-        margin-bottom: theme.spacing(24);
-        width: 100%;
-        height: theme.spacing(80);
-      }
-
-      .footerSocialLinks {
-        gap: theme.spacing(8);
-        justify-content: center;
-        margin-bottom: theme.spacing(6);
-        margin-left: -#{theme.spacing(1)};
-        margin-right: -#{theme.spacing(1)};
-
-        @include theme.breakpoint("lg") {
-          display: none;
-        }
-      }
-
-      .legal {
-        @include theme.text("xs", "normal");
-
-        color: theme.color("muted");
-        display: flex;
-        flex-flow: row nowrap;
-        gap: theme.spacing(8);
-
-        .footerLink {
-          color: theme.color("muted");
-        }
-      }
-
-      @include theme.breakpoint("md") {
-        display: grid;
-        grid-template-columns: repeat(2, minmax(max-content, 1fr));
-        grid-template-rows: repeat(2, max-content);
-        align-items: center;
-
-        .lines {
-          grid-column: span 2 / span 2;
-        }
-
-        .footerSocialLinks {
-          grid-row: 2;
-          grid-column: 2;
-          margin-bottom: 0;
-          justify-content: flex-end;
-        }
-
-        .legal {
-          grid-row: 2;
-          grid-column: 1;
-          text-align: left;
-          padding: theme.spacing(2) 0;
-        }
-      }
-    }
-  }
-
-  // We also have to disable `scrollbar-gutter: stable` when overlays are
-  // visible, because chrome leaves an unsightly gap rather than letting the
-  // modal backgrop fill the page even though it's fixed position.
-  &[data-overlay-visible] {
-    scrollbar-gutter: auto;
-    overflow: hidden;
-
-    .body {
-      padding-right: var(--scrollbar-width);
-    }
-  }
+.root,
+.body {
+  width: 100%;
+  height: 100%;
 }
