+++ apps/portfolio/src/components/Root/index.tsx
@@ -1,15 +1,18 @@
+import { FogoSessionProvider } from "@fogo/sessions-sdk-react";
 import { GoogleAnalytics } from "@next/third-parties/google";
 import clsx from "clsx";
 import dynamic from "next/dynamic";
 import { Funnel_Display } from "next/font/google";
 import type { ReactNode } from "react";
 
-import "./root.scss";
 import styles from "./index.module.scss";
+import "./root.scss";
 import { RouterProvider } from "./router-provider";
 import {
   ENABLE_ACCESSIBILITY_REPORTING,
   GOOGLE_ANALYTICS_ID,
+  DOMAIN,
+  RPC,
 } from "../../config/server";
 import { LoggerProvider } from "../../hooks/use-logger";
 
@@ -30,7 +33,11 @@ export const Root = ({ children }: Props) => (
   <RouterProvider>
     <LoggerProvider>
       <html lang="en" className={clsx(sans.className, styles.root)}>
-        <body className={styles.body}>{children}</body>
+        <body className={styles.body}>
+          <FogoSessionProvider endpoint={RPC} domain={DOMAIN}>
+            {children}
+          </FogoSessionProvider>
+        </body>
         {GOOGLE_ANALYTICS_ID && <GoogleAnalytics gaId={GOOGLE_ANALYTICS_ID} />}
         {ENABLE_ACCESSIBILITY_REPORTING && <ReportAccessibility />}
       </html>
