+++ apps/portfolio/vercel.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://openapi.vercel.sh/vercel.json",
+  "buildCommand": "./build.sh",
+  "installCommand": "./install.sh"
+}
