+++ packages/sessions-sdk-react/src/components/session-button.module.scss
@@ -174,118 +174,18 @@
         opacity 200ms;
     }
 
-    .sessionPanel {
+    .sessionPanelDialog {
       width: 100%;
       height: theme.spacing(148);
       max-height: 80dvh;
-      border-top-left-radius: theme.border-radius("xl");
-      border-top-right-radius: theme.border-radius("xl");
-      background: theme.color("card");
-      border: 1px solid theme.color("widget-border");
-      box-shadow: theme.shadow();
-      display: grid;
-      grid-template-rows: theme.spacing(12) 1fr theme.spacing(12);
       outline: none;
 
       @include theme.breakpoint("sm") {
         width: theme.spacing(88);
-        border-radius: theme.border-radius("xl");
       }
 
-      .header {
-        border-bottom: 1px solid theme.color("border");
-        display: flex;
-        flex-flow: row nowrap;
-        padding: theme.spacing(2);
-        padding-left: theme.spacing(4);
-        gap: theme.spacing(2);
-        align-items: center;
-
-        .title {
-          @include theme.text("sm", "medium");
-
-          color: theme.color("heading");
-          flex-grow: 1;
-        }
-      }
-
-      .tabs {
-        display: grid;
-        grid-template-rows: theme.spacing(10) 1fr;
-        overflow: hidden;
-
-        .tabList {
-          display: flex;
-          flex-flow: row nowrap;
-          gap: theme.spacing(1);
-          border-bottom: 1px solid theme.color("widget-border");
-
-          .tab {
-            @include theme.text("xs", "semibold");
-
-            color: theme.color("muted");
-            flex-grow: 1;
-            display: grid;
-            place-content: center;
-            position: relative;
-            outline: 2px solid transparent;
-            outline-offset: -#{theme.spacing(0.5)};
-            transition:
-              color 50ms linear,
-              background-color 50ms linear,
-              outline-color 50ms linear;
-
-            .underline {
-              position: absolute;
-              bottom: 0;
-              left: theme.spacing(3);
-              right: theme.spacing(3);
-              height: theme.spacing(0.5);
-              background: theme.color("accent");
-              border-top-left-radius: theme.border-radius("sm");
-              border-top-right-radius: theme.border-radius("sm");
-            }
-
-            &:not([data-selected]) {
-              cursor: pointer;
-
-              &[data-hovered] {
-                background: theme.color("button-hover");
-                color: theme.color("foreground");
-              }
-
-              &[data-pressed] {
-                background: theme.color("button-pressed");
-                color: theme.color("foreground");
-              }
-            }
-
-            &[data-focus-visible] {
-              outline-color: theme.color("accent");
-            }
-          }
-        }
-      }
-
-      .tabPanel {
-        overflow: hidden;
-      }
-
-      .footer {
-        border-top: 1px solid theme.color("border");
-        background: theme.color("demo-bg");
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        padding: theme.spacing(2);
-        padding-left: theme.spacing(4);
-        border-bottom-left-radius: theme.border-radius("xl");
-        border-bottom-right-radius: theme.border-radius("xl");
-
-        .fogoWordmark {
-          height: theme.spacing(5);
-          color: theme.pallette-color("white");
-        }
+      .sessionPanel {
+        box-shadow: theme.shadow();
       }
     }
