+++ packages/sessions-sdk-react/src/index.tsx
@@ -8,3 +8,4 @@ export {
 export { useSession } from "./hooks/use-session.js";
 export { FogoSessionProvider } from "./components/session-provider.js";
 export { SessionButton } from "./components/session-button.js";
+export { SessionPanel } from "./components/session-panel.js";
