+++ packages/sessions-sdk-react/src/theme.scss
@@ -549,6 +549,7 @@ $shadow: (
     font-style: normal;
     color-scheme: dark;
     color: purple;
+    text-align: left;
 
     *::selection {
       color: pallette-color("black");
